Update spec and tests
Update sass-spec to the 'libsass-3.6.3' tag.

Add a pretest script for copying the sass-spec module to a working
directory and for expanding the hrx archive files into individual files.

Update spec.js to account for changes in the spec layout.

Update test fixtures to account for changes in spec requirements.